Use the correct Ruby versioning
Instead of 'inventing' MoJ specific Ruby versioning, how about this
crazy idea: we use the same versioning for Ruby as Ruby itself?

Mental, I know, but I suspect less confusing than the current state.

So when the next version of Ruby comes out, we follow this example and
we create a new Docker image based on the version of Ruby. This will
allow developers to use the exact Ruby version their app needs without
upgrading the images for all the others in MoJ:

e.g. Ruby 2.2.1 is not the same as Ruby 2.2.4

But this is what was happening until now.